Update config poller to not take a context parameter for watch
Motivation
----------
The context parameter that we pass into the config poller at the
moment is the same context passed into CreateAgent. This context
shoould only be applied for initial setup of the agent, not
the lifetime. Creating a separate context and storing it on agent
would be against go idioms so we should also add a Close function
to the poller.