ING-588: Refactor buckets tracking agent manager
Motivation
-----------
At the moment the buckets tracking agent manager will never ever
return a bucket not found error for a GetAgent call. We need
to rework the manager so that it returns bucket not found if the
bucket does not exist on the cluster, rather than waiting until
context is errored.

Changes
-------
Refactor a buckets tracking watcher out of the buckets tracking
agent manager.
Update bucket tracking watcher so that when GetAgent is called
for an unknown bucket then it triggers the watcher to attempt
a manual update.
Update watcher logic so that if the request to GetAgent was made
after a manual update was already in flight then it will trigger
another manual update if the bucket is still unknown afterward.