ING-572: Add a fast map to buckets tracking http
Motivation
----------
At the moment if a user calls GetAgent when the tracker is handling
detection of new buckets then the call will always block until
all new agents are created and old ones destroyed. We should add
a fast map so that requests for agents that we already know about
can be resolved without waiting for a global lock.

Changes
-------
Replace the agents map in the bucket tracker watcher with 2 new
maps - a fast map and a slow map.
When we handle buckets build both the slow map and the fast map.
In GetAgent consult the fast map first and if the bucket is not
present then trigger an update to fetch all buckets and update
the maps.