ING-590: Return invalid args error on write packet validation fail
Motivation
----------
On the write side of writing to the connection anything which
causes the validation of the packet to fail should be an invalid
args error rather than a protocol error. It means that there is
an issue in the user's packet rather than a protocol issue like
there would be for a response packet.

Changes
-------
Updated WritePacket to return invalid argument errors on packet
validation failure.