ING-523: Improve GetAllReplicas handling
The logic here starts a thread for each replica that should
exist according to NUM_REPLICAS. The overall GetAllReplicas
operation is complete when NUM_REPLICAS + 1 docs have been
returned, or it hits the timeout. The individual threads
will run indefinitely until these points, even if the thread
has returned a document. This is because the individual
requests are routed based on the replica ID, which can
be routed to a different node if there is a rebalance during
the GetAllReplicas.