ING-523: Improve GetAllReplicas handling
Here the logic tries to get the maximum possible number of
replicas and the original document in parallel, streaming
any results/errors returned to the User with two exceptions. We
only return one document per node, to avoid changind replica Ids
causing the same replica to be returned twice. And we ignore
ErrInvalidReplica, since we naievely try and get the max possible
number of replicas we expect to see this.